requirements.txt
python-dotenv
pymongo
openai
pdfplumber
nltk
PyPDF2
pyyaml
